sum of floats

+++ python_variable_annotations/7-to_kv.py
@@ -7,7 +7,7 @@ Import:
 from typing import Union, Tuple
 
 
-def to_kv(k: str, v: Union[int, float]) -> Tuple[str, int]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """Function takes to args and returns them in tuple
 
     Args:
